Clean code anpassungen, Entfernen des Consolen output da nichtmehr nötig dank Gui, Anpassungen in diversen systemen
